# docs/openapi.json
{
  "components" : {
    "schemas" : {
      "Actual" : {
        "properties" : {
          "breakdown" : {
            "$ref" : "#/components/schemas/FareBreakdownResponse"
          },
          "distanceMeters" : {
            "format" : "int32",
            "type" : "integer"
          },
          "distanceSource" : {
            "enum" : [ "TRACKED", "ESTIMATED" ],
            "type" : "string"
          },
          "durationSeconds" : {
            "format" : "int32",
            "type" : "integer"
          },
          "fare" : {
            "$ref" : "#/components/schemas/Money"
          }
        },
        "required" : [ "breakdown", "distanceMeters", "distanceSource", "durationSeconds", "fare" ],
        "type" : "object"
      },
...
      "Cancellation" : {
        "properties" : {
          "cancelledBy" : {
            "enum" : [ "PASSENGER", "DRIVER", "SYSTEM", "ADMIN" ],
            "type" : "string"
          },
          "reason" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "cancelledBy" ],
        "type" : "object"
      },
...
      "DriverDecisionRequest" : {
        "properties" : {
          "reason" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "reason" ],
        "type" : "object"
      },
      "DriverInfo" : {
        "properties" : {
          "fullName" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "ratingAvg" : {
            "type" : [ "number", "null" ]
          },
          "ratingCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "vehicle" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/Vehicle"
            }, {
              "type" : "null"
            } ]
          }
        },
        "required" : [ "fullName", "id", "ratingCount" ],
        "type" : "object"
      },
...
      "DriverResponse" : {
        "properties" : {
          "availability" : {
            "enum" : [ "OFFLINE", "AVAILABLE", "ON_TRIP" ],
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "fullName" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "licenseNumber" : {
            "type" : "string"
          },
          "ratingAvg" : {
            "type" : [ "number", "null" ]
          },
          "ratingCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "rejectionReason" : {
            "type" : [ "string", "null" ]
          },
          "vehicle" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/VehicleResponse"
            }, {
              "type" : "null"
            } ]
          },
          "verificationStatus" : {
            "enum" : [ "PENDING", "VERIFIED", "REJECTED", "SUSPENDED" ],
            "type" : "string"
          },
          "verifiedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "availability", "createdAt", "email", "fullName", "id", "licenseNumber", "ratingCount", "verificationStatus" ],
        "type" : "object"
      },
...
      "Estimate" : {
        "properties" : {
          "breakdown" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/FareBreakdownResponse"
            }, {
              "type" : "null"
            } ]
          },
          "distanceMeters" : {
            "format" : "int32",
            "type" : "integer"
          },
          "durationSeconds" : {
            "format" : "int32",
            "type" : "integer"
          },
          "fare" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/Money"
            }, {
              "type" : "null"
            } ]
          },
          "source" : {
            "enum" : [ "ROUTED", "APPROXIMATE" ],
            "type" : "string"
          }
        },
        "required" : [ "distanceMeters", "durationSeconds", "source" ],
        "type" : "object"
      },
...
      "GeoPoint" : {
        "properties" : {
          "lat" : {
            "format" : "double",
            "maximum" : 90.0,
            "minimum" : -90.0,
            "type" : "number"
          },
          "lng" : {
            "format" : "double",
            "maximum" : 180.0,
            "minimum" : -180.0,
            "type" : "number"
          }
        },
        "required" : [ "lat", "lng" ],
        "type" : "object"
      },
...
      "Matching" : {
        "properties" : {
          "radiusMeters" : {
            "format" : "int32",
            "type" : "integer"
          },
          "round" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "radiusMeters", "round" ],
        "type" : "object"
      },
      "Money" : {
        "properties" : {
          "amount" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          }
        },
        "required" : [ "amount", "currency" ],
        "type" : "object"
      },
...
      "NotificationResponse" : {
        "properties" : {
          "body" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "read" : {
            "type" : "boolean"
          },
          "rideId" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "enum" : [ "DRIVER_ACCEPTED", "DRIVER_ARRIVING", "DRIVER_ARRIVED", "DRIVER_REASSIGNING", "TRIP_STARTED", "TRIP_COMPLETED", "RIDE_CANCELLED", "RIDE_EXPIRED", "PAYMENT_RECEIVED", "EARNINGS_RECORDED", "DRIVER_VERIFIED", "DRIVER_REJECTED" ],
            "type" : "string"
          }
        },
        "required" : [ "body", "createdAt", "id", "read", "title", "type" ],
        "type" : "object"
      },
...
      "PageResponseNotificationResponse" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/NotificationResponse"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sort" : {
            "type" : "string"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "content", "page", "size", "sort", "totalElements", "totalPages" ],
        "type" : "object"
      },
...
      "PassengerInfo" : {
        "properties" : {
          "fullName" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "fullName", "id" ],
        "type" : "object"
      },
      "PaymentInfo" : {
        "properties" : {
          "amount" : {
            "$ref" : "#/components/schemas/Money"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "method" : {
            "enum" : [ "CASH", "CARD" ],
            "type" : "string"
          },
          "provider" : {
            "enum" : [ "CASH", "SANDBOX" ],
            "type" : "string"
          },
          "status" : {
            "enum" : [ "PENDING", "CAPTURED", "FAILED", "REFUNDED" ],
            "type" : "string"
          }
        },
        "required" : [ "amount", "id", "method", "provider", "status" ],
        "type" : "object"
      },
...
      "PlaceResponse" : {
        "properties" : {
          "address" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "point" : {
            "$ref" : "#/components/schemas/GeoPoint"
          }
        },
        "required" : [ "address", "name", "point" ],
        "type" : "object"
      },
...
      "RidePlace" : {
        "properties" : {
          "address" : {
            "type" : "string"
          },
          "point" : {
            "$ref" : "#/components/schemas/GeoPoint"
          }
        },
        "required" : [ "address", "point" ],
        "type" : "object"
      },
      "RideResponse" : {
        "properties" : {
          "actual" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/Actual"
            }, {
              "type" : "null"
            } ]
          },
          "cancellation" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/Cancellation"
            }, {
              "type" : "null"
            } ]
          },
          "driver" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/DriverInfo"
            }, {
              "type" : "null"
            } ]
          },
          "dropoff" : {
            "$ref" : "#/components/schemas/RidePlace"
          },
          "estimate" : {
            "$ref" : "#/components/schemas/Estimate"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "matching" : {
            "$ref" : "#/components/schemas/Matching"
          },
          "passenger" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/PassengerInfo"
            }, {
              "type" : "null"
            } ]
          },
          "payment" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/PaymentInfo"
            }, {
              "type" : "null"
            } ]
          },
          "paymentMethod" : {
            "enum" : [ "CASH", "CARD" ],
            "type" : "string"
          },
          "pickup" : {
            "$ref" : "#/components/schemas/RidePlace"
          },
          "status" : {
            "enum" : [ "REQUESTED", "MATCHING", "DRIVER_ASSIGNED", "DRIVER_ARRIVING", "DRIVER_ARRIVED", "IN_PROGRESS", "COMPLETED", "CANCELLED", "EXPIRED" ],
            "type" : "string"
          },
          "timestamps" : {
            "$ref" : "#/components/schemas/Timestamps"
          },
          "vehicleCategory" : {
            "enum" : [ "ECONOMY", "COMFORT", "XL" ],
            "type" : "string"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "dropoff", "estimate", "id", "matching", "paymentMethod", "pickup", "status", "timestamps", "vehicleCategory", "version" ],
        "type" : "object"
      },
...
      "RouteResponse" : {
        "properties" : {
          "distanceMeters" : {
            "format" : "int32",
            "type" : "integer"
          },
          "durationSeconds" : {
            "format" : "int32",
            "type" : "integer"
          },
          "path" : {
            "items" : {
              "$ref" : "#/components/schemas/GeoPoint"
            },
            "type" : "array"
          },
          "source" : {
            "enum" : [ "ROUTED", "APPROXIMATE" ],
            "type" : "string"
          }
        },
        "required" : [ "distanceMeters", "durationSeconds", "path", "source" ],
        "type" : "object"
      },
...
      "Timestamps" : {
        "properties" : {
          "acceptedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "arrivedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "cancelledAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "completedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "enRouteAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "expiredAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "requestedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "startedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "requestedAt" ],
        "type" : "object"
      },
...
      "VehicleResponse" : {
        "properties" : {
          "category" : {
            "enum" : [ "ECONOMY", "COMFORT", "XL" ],
            "type" : "string"
          },
          "color" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "make" : {
            "type" : "string"
          },
          "model" : {
            "type" : "string"
          },
          "modelYear" : {
            "format" : "int32",
            "type" : "integer"
          },
          "plateNumber" : {
            "type" : "string"
          },
          "seats" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "category", "color", "id", "make", "model", "modelYear", "plateNumber", "seats" ],
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Real-time ride-hailing platform API. Conventions: docs/api.md",
    "title" : "RideFlow API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/admin/drivers/{driverId}/reject" : {
      "post" : {
        "operationId" : "reject_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "driverId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DriverDecisionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DriverResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Reject a pending driver application",
        "tags" : [ "Admin: drivers" ]
      }
    },
...
    "/api/geo/reverse" : {
      "get" : {
        "operationId" : "reverse",
        "parameters" : [ {
          "in" : "query",
          "name" : "lat",
          "required" : true,
          "schema" : {
            "format" : "double",
            "maximum" : 90.0,
            "minimum" : -90.0,
            "type" : "number"
          }
        }, {
          "in" : "query",
          "name" : "lng",
          "required" : true,
          "schema" : {
            "format" : "double",
            "maximum" : 180.0,
            "minimum" : -180.0,
            "type" : "number"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlaceResponse"
                }
              }
            },
            "description" : "The address"
          },
          "204" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlaceResponse"
                }
              }
            },
            "description" : "No address at this point"
          }
        },
        "summary" : "Address at a point, or 204 if there is none (for example at sea)",
        "tags" : [ "Geo" ]
      }
    },
    "/api/geo/route" : {
      "get" : {
        "operationId" : "route",
        "parameters" : [ {
          "in" : "query",
          "name" : "fromLat",
          "required" : true,
          "schema" : {
            "format" : "double",
            "maximum" : 90.0,
            "minimum" : -90.0,
            "type" : "number"
          }
        }, {
          "in" : "query",
          "name" : "fromLng",
          "required" : true,
          "schema" : {
            "format" : "double",
            "maximum" : 180.0,
            "minimum" : -180.0,
            "type" : "number"
          }
        }, {
          "in" : "query",
          "name" : "toLat",
          "required" : true,
          "schema" : {
            "format" : "double",
            "maximum" : 90.0,
            "minimum" : -90.0,
            "type" : "number"
          }
        }, {
          "in" : "query",
          "name" : "toLng",
          "required" : true,
          "schema" : {
            "format" : "double",
            "maximum" : 180.0,
            "minimum" : -180.0,
            "type" : "number"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RouteResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Road route between two points (straight-line fallback flagged APPROXIMATE)",
        "tags" : [ "Geo" ]
      }
    },
    "/api/geo/search" : {
      "get" : {
        "operationId" : "search",
        "parameters" : [ {
          "in" : "query",
          "name" : "q",
          "required" : true,
          "schema" : {
            "maxLength" : 200,
            "minLength" : 2,
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "lat",
          "required" : false,
          "schema" : {
            "format" : "double",
            "maximum" : 90.0,
            "minimum" : -90.0,
            "type" : "number"
          }
        }, {
          "in" : "query",
          "name" : "lng",
          "required" : false,
          "schema" : {
            "format" : "double",
            "maximum" : 180.0,
            "minimum" : -180.0,
            "type" : "number"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/PlaceResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Place search, biased towards lat/lng when given; call on submit, not per keystroke",
        "tags" : [ "Geo" ]
      }
    },
    "/api/notifications" : {
      "get" : {
        "operationId" : "list",
        "parameters" : [ {
          "in" : "query",
          "name" : "unreadOnly",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "maximum" : 100,
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "default" : "createdAt,desc",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseNotificationResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Newest first by default",
        "tags" : [ "Notifications" ]
      }
    },
...
    "/api/notifications/{notificationId}/read" : {
      "post" : {
        "operationId" : "markRead",
        "parameters" : [ {
          "in" : "path",
          "name" : "notificationId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "Notifications" ]
      }
    },
...
    "/api/rides/{rideId}" : {
      "get" : {
        "operationId" : "get",
        "parameters" : [ {
          "in" : "path",
          "name" : "rideId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RideResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Rides" ]
      }
    },
    "/api/rides/{rideId}/accept" : {
      "post" : {
        "operationId" : "accept",
        "parameters" : [ {
          "in" : "path",
          "name" : "rideId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RideResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Accept an open offer (first driver to accept wins)",
        "tags" : [ "Rides: driver actions" ]
      }
    },
    "/api/rides/{rideId}/arrive" : {
      "post" : {
        "operationId" : "arrive",
        "parameters" : [ {
          "in" : "path",
          "name" : "rideId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RideResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Arrived at pickup (requires a fresh location inside the pickup geofence)",
        "tags" : [ "Rides: driver actions" ]
      }
    },
...
    "/api/rides/{rideId}/complete" : {
      "post" : {
        "operationId" : "complete",
        "parameters" : [ {
          "in" : "path",
          "name" : "rideId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RideResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Complete the trip; measures distance from the GPS trail and computes the final fare",
        "tags" : [ "Rides: driver actions" ]
      }
    },
    "/api/rides/{rideId}/en-route" : {
      "post" : {
        "operationId" : "enRoute",
        "parameters" : [ {
          "in" : "path",
          "name" : "rideId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RideResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Start driving to the pickup",
        "tags" : [ "Rides: driver actions" ]
      }
    },
...
    "/api/rides/{rideId}/reject" : {
      "post" : {
        "operationId" : "reject",
        "parameters" : [ {
          "in" : "path",
          "name" : "rideId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "summary" : "Decline an open offer",
        "tags" : [ "Rides: driver actions" ]
      }
    },
    "/api/rides/{rideId}/start" : {
      "post" : {
        "operationId" : "start",
        "parameters" : [ {
          "in" : "path",
          "name" : "rideId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RideResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Passenger on board; the trip starts",
        "tags" : [ "Rides: driver actions" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "url" : "/"
  } ]
}